Keygen update
I also added the RequestId to the `KeygenResponse.FromBinary`, and I use this function here so that removes two levels.
I also removed the try catch, as there shouldn't be anything unexpected, and the whole function is in a try catch already.
Finally, I removed the `message.Contains` as performance shouldn't be an issue with this type of messages that are supposed to be sent scarcely. Of course, I check RequestId is not null.

The resulting code has 3 less levels. I think it's cleaner.