C# sdk V2 (#13)
* Options refac

* Presence WIP

* Presence WIP

* Reorg with partial classes

* Trie generalisation

Todos to check

* Presence WIP

* Presence

* Update PresenceEvent.cs

Simplification

* Error event

some work

* Me WIP

* PresenceEvent requestID long

* Link WIP

* Added Benchmarks project

* Benchmark trim

* Update Bench Trim

* Format channel refact

* Link

* Sample WIP

* Me handler in Me.cs

* Subscribe refac WIP

* Update Net40

Add Net471

* PresenceEvent.cs: tiny code improvement, casting with is

* SubscribeWithGroup

* Me links

* sample update

* Sample update

* Added missing files to all .net versions

* Added TLS connection

* Update sample

* Subscribe WithFrom WithUntil

* Fix match trie + unit test

* More Trie tests

* Added async KeyGen function (#11)

* Added async KeyGen function

* Made cosmetic changes

* WinRT compat updates

* Update Emitter.cs

* Keygen update

I also added the RequestId to the `KeygenResponse.FromBinary`, and I use this function here so that removes two levels.
I also removed the try catch, as there shouldn't be anything unexpected, and the whole function is in a try catch already.
Finally, I removed the `message.Contains` as performance shouldn't be an issue with this type of messages that are supposed to be sent scarcely. Of course, I check RequestId is not null.

The resulting code has 3 less levels. I think it's cleaner.

* ToTimestamps handle non utc time

* .net 4.6.2 project

* .net 4.7 project

* .net 4.7.1 project

* .net 4.7.2 project

* Removed async keygen (for now) and cleaned sample

* WIP documentation

* Presence and keygen error throwing

* Documentation

* Fix presence handler and optional handler